+++ tests/Nexucis/Tests/Elasticsearch/Helper/Nodowntime/DocumentActionTest.php
@@ -2,7 +2,7 @@
 
 namespace Nexucis\Tests\Elasticsearch\Helper\Nodowntime;
 
-use Nexucis\Elasticsearch\Helper\Nodowntime\Parameter\SearchParameter;
+use Nexucis\Elasticsearch\Helper\Nodowntime\DumpNamespace\SearchParameter;
 use stdClass;
 
 class DocumentActionTest extends AbstractIndexHelperTest
